move mock data into a separate file
